requirements.txt
Flask
Flask-Bootstrap
Flask-FontAwesome
Flask-Session
Flask-SQLAlchemy
Flask-WTF
Flask-Script
Flask-Mobility
Flask-Migrate
Flask-JWT-Extended
Jinja2
mysql-connector-python
python-dateutil
python-dotenv
SQLAlchemy
dictalchemy
Werkzeug
WTForms
requests
cachetools
googletrans
email_validator
html2text
pyotp
phonenumbers
twilio
ipinfo
whatsmyip
